requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
anthropic>=0.28.0
httpx>=0.27.0
pyyaml>=6.0.1
jsonschema>=4.22.0
python-multipart>=0.0.9
pypdf>=4.0.0
pytest-httpserver>=1.1.0
